Conflicts:
	cmake/DefaultExecutable.cmake
	cmake/DefaultLibrary.cmake
	cmake/DefaultProject.cmake
	src/executables/IO_Test/PlaceHolderListeners.cpp
	src/executables/IO_Test/main.cpp
	src/executables/Resource_Test/main.cpp

+++ src/libraries/Physics/PhysicWorld.cpp
@@ -66,7 +66,7 @@ void PhysicWorld::screenPosToWorldRay(double mouseX, double mouseY, int screenWi
 	PhysicWorld::getInstance()->dynamicsWorld->rayTest(btVector3(outOrigin.x, outOrigin.y, outOrigin.z), btVector3(outDirection.x, outDirection.y, outDirection.z), RayCallback);
 
 	if(RayCallback.hasHit()) {
-	    cout << "mesh: " << (int)RayCallback.m_collisionObject->getUserPointer() << endl;
+	    cout << "mesh: " << RayCallback.m_collisionObject->getUserPointer() << endl;
 	}
 	else{
 	    cout << "background" << endl;
